CLean up

+++ Server/Network/Packets.cs
@@ -3044,7 +3044,7 @@ namespace Server.Network
 			m_Stream.Write((short)m.ManaMax);
 
 			m_Stream.Write(m.TotalGold);
-			m_Stream.Write((short)(Core.AOS ? m.PhysicalResistance : (int)(m.ArmorRating + 0.5)));
+			m_Stream.Write((short)(Core.AOS ? m.PhysicalResistance : m.ArmorRating + 0.5));
 			m_Stream.Write((short)(Mobile.BodyWeight + m.TotalWeight));
 
 			if (type >= 5)
@@ -3146,7 +3146,7 @@ namespace Server.Network
 				WriteAttr(beheld.Mana, beheld.ManaMax);
 
 				m_Stream.Write(beheld.TotalGold);
-				m_Stream.Write((short)(Core.AOS ? beheld.PhysicalResistance : (int)(beheld.ArmorRating + 0.5)));
+				m_Stream.Write((short)(Core.AOS ? beheld.PhysicalResistance : beheld.ArmorRating + 0.5));
 				m_Stream.Write((short)(Mobile.BodyWeight + beheld.TotalWeight));
 
 				if (type >= 5)
